+++ auth-common/src/main/java/net/therap/auth/client/TokenClient.java
@@ -1,5 +1,6 @@
 package net.therap.auth.client;
 
+import net.therap.auth.discovery.ServiceDiscoveryCacheAuth;
 import net.therap.auth.dto.RefreshTokenRequestDto;
 import net.therap.auth.dto.TokenResponseDto;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,12 +24,12 @@ public class TokenClient {
     private static final String REFRESH_PATH = "/auth/api/refresh";
 
     @Autowired
-    private ServiceDiscoveryCacheForAuthCommon serviceDiscoveryCacheForAuthCommon;
+    private ServiceDiscoveryCacheAuth serviceDiscoveryCacheAuth;
 
     private final RestTemplate restTemplate = new RestTemplate();
 
     private String getServiceBaseUrl() {
-        Map<String, Object> instance = serviceDiscoveryCacheForAuthCommon.getInstance(SERVICE_NAME);
+        Map<String, Object> instance = serviceDiscoveryCacheAuth.getInstance(SERVICE_NAME);
         String host = (String) instance.get("host");
         Integer port = (Integer) instance.get("port");
         return "http:" + host + ":" + port;
