--- auth-common/src/main/resources/application.properties
@@ -1,24 +1,0 @@
-spring.application.name=LearnHub
-server.port=8080
-service.name=auth
-service.host=127.0.0.1
-registry.url=http://192.168.0.215:8761
-
-logging.level.org.springframework.security=DEBUG
-feign.httpclient.disable-ssl-validation=true
-
-spring.mail.host=smtp.gmail.com
-spring.mail.port=587
-spring.mail.username=apurbo.turjo@therapservices.net
-spring.mail.password=yeoe ijka gtsn ztzd
-spring.mail.properties.mail.smtp.auth=true
-spring.mail.properties.mail.smtp.starttls.enable=true
-
-spring.security.user.name=user
-spring.security.user.password=test
-
-token.secret.key=secretKeyTest123
-
-jst.key-id.default-key=therap
-jwt.private-key-path=key/private_key.pem
-jwt.public-key-path=key/public_key.pem
