+++ authserver/src/main/java/net/therap/service/VerificationTokenService.java
@@ -5,7 +5,6 @@ import lombok.RequiredArgsConstructor;
 import net.therap.entity.User;
 import net.therap.entity.VerificationToken;
 import net.therap.respository.VerificationTokenRepository;
-import net.therap.service.interfaces.EmailService;
 import org.springframework.stereotype.Service;
 
 import java.util.UUID;
