--- authserver/src/main/java/net/therap/service/interfaces/EmailService.java
@@ -1,10 +1,0 @@
-package net.therap.service.interfaces;
-
-/**
- * @author apurboturjo
- * @since 7/29/25
- */
-public interface EmailService {
-    
-    void sendVerificationEmail(String to, String token);
-}
