+++ authserver/src/main/resources/application.properties
@@ -12,7 +12,7 @@ spring.datasource.url=jdbc:oracle:thin:@bddb02.therapbd.net:1521/trainee_pdb.bdd
 spring.datasource.username=trainee2
 spring.datasource.password=trainee2
 spring.datasource.driver-class-name=oracle.jdbc.OracleDriver
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=none
 spring.jpa.show-sql=true
 spring.datasource.hikari.auto-commit=false
 spring.datasource.hikari.minimum-idle=5
