--- authserver/src/main/resources/migration/DDL.sql
@@ -1,7 +1,0 @@
--- DDL Script for employee table
-CREATE TABLE employee (
-    id INT PRIMARY KEY,
-    email VARCHAR(255) UNIQUE NOT NULL,
-    firstname VARCHAR(100) NOT NULL,
-    lastname VARCHAR(100) NOT NULL
-);
